Reorganize test with environment variables.
Environment variables make it much easier to test the matrix of
conditions, as well as roll out new tests that all operate with the same
procedure.

Additionally, I've removed the custom my.cnf as that is no longer
necessary. I tried using it before to allow for bigger tables, but that
issue was fixed when I reduced the table size for the tests.